Rename kubeclean module to resemble command
